Renaming lbp to mpr
